refactor: 상수 타입들 파일로 분리
