+++ StreamChat/StreamChat/Source/Constant/MessageConstant.swift
@@ -1,0 +1,16 @@
+//
+//  MessageConstant.swift
+//  StreamChat
+//
+//  Created by steven on 8/13/21.
+//
+
+import Foundation
+
+enum MessageConstant {
+    static let userPrefix = "USR_NAME"
+    static let sendPrefix = "MSG"
+    static let exitPrefix = "LEAVE"
+    static let separator = "::"
+    static let postfix = "END"
+}
