+++ StreamChat/StreamChat/Source/Constant/MessageType.swift
@@ -1,0 +1,14 @@
+//
+//  MessageType.swift
+//  StreamChat
+//
+//  Created by steven on 8/13/21.
+//
+
+import Foundation
+
+enum MessageType {
+    case me
+    case other
+    case system
+}
