+++ StreamChat/StreamChat/Source/Model/Message.swift
@@ -7,50 +7,6 @@
 
 import Foundation
 
-enum MessageConstant {
-    static let userPrefix = "USR_NAME"
-    static let sendPrefix = "MSG"
-    static let exitPrefix = "LEAVE"
-    static let separator = "::"
-    static let postfix = "END"
-}
-
-enum OutgoingMessage {
-    case enter(String)
-    case send(String)
-    case leave
-}
-
-extension OutgoingMessage: CustomStringConvertible {
-    var description: String {
-        switch self {
-        case .enter(let user):
-            return MessageConstant.userPrefix +
-                    MessageConstant.separator +
-                    user +
-                    MessageConstant.separator +
-                    MessageConstant.postfix
-        case .send(let message):
-            return MessageConstant.sendPrefix +
-                    MessageConstant.separator +
-                    message +
-                    MessageConstant.separator +
-                    MessageConstant.postfix
-        case .leave:
-            return MessageConstant.exitPrefix +
-                    MessageConstant.separator +
-                    MessageConstant.separator +
-                    MessageConstant.postfix
-        }
-    }
-}
-
-enum MessageType {
-    case me
-    case other
-    case system
-}
-
 struct Message {
     let username: String
     let content: String
